capy_board_init: remove reponsibilities of the top level
To prevent libcapybara becoming kitchen sink, its scope should
be limited to (reconfigurable) power system, not common code
for a particular board. Also, unlocking GPIOs, setting up
system clock, initializing buses, etc is not the responsibility of any
library -- this is code that should be owned by main(), because
multiple libraries are affected by this code. The plan is to
move more stuff out of libcapybara in separate commits.

If it is desirable to put some of this initialization code that's common
to a specific board (e.g. the Capybara board), in order to avoid
duplicating code across apps, then a new library for this very purpose
should be created. Libcapybara has a different purpose.